Add files via upload
These are the HTML Files that I have to cycle through with my spark functions. However I've been having some difficulty making them work properly. I have to get the entries from the prof and enter it into the answer question for the students. I can get the information using GET and I can post it using POST. I just have to get the variables to work together. In each html file there is a form action, that hasnt been updated I just copied it from the lab for now. I have to change them to the functions that I have going in my spark program. I'm going to come home after work tomorrow and work on the spark some more. When testing the html files just make sure logo and UOIT are in the same folder. Let me know what to add to the UI as well. I didn't want it to look to junky.